backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0
python-dotenv
openai
requests
moviepy
SpeechRecognition
aiofiles
jinja2
pydantic-settings
pydub
beautifulsoup4
aiohttp
python-slugify
langdetect
reportlab
